# openapi.yaml
openapi: "3.0.0"
info:
    title: WhatsApp API MultiDevice
    version: 8.2.0
    description: |
        This API is used for sending whatsapp via API.

        Device scoping:
        - Send `X-Device-Id` on all device-scoped REST calls.
        - WebSocket: connect to `/ws?device_id=<id>`.
servers:
    - url: http://localhost:3000
...
security:
    - basicAuth: []
paths:
...
    /send/message:
        post:
            operationId: sendMessage
            tags:
                - send
            summary: Send Message
            parameters:
                - $ref: "#/components/parameters/DeviceIdHeader"
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                phone:
                                    type: string
                                    example: "6289685028129@s.whatsapp.net"
                                    description: Phone number with country code
                                message:
                                    type: string
                                    example: selamat malam
                                    description: Message to send
                                reply_message_id:
                                    type: string
                                    example: 3EB089B9D6ADD58153C561
                                    description: Message ID that you want reply
                                is_forwarded:
                                    type: boolean
                                    example: false
                                    description: Whether this is a forwarded message
                                duration:
                                    type: integer
                                    example: 3600
                                    description: Disappearing message duration in seconds (optional)
                                mentions:
                                    type: array
                                    items:
                                        type: string
                                    example: ["628123456789", "@everyone"]
                                    description: |
                                        List of phone numbers to mention (ghost mentions - no @ required in message text).
                                        Use special keyword "@everyone" to mention all group participants.
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/SendResponse"
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorBadRequest"
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorInternalServer"
...
components:
    parameters:
        DeviceIdHeader:
            name: X-Device-Id
            in: header
            required: false
            description: |
                Device identifier for multi-device support. Required when multiple devices are registered.
                If only one device is registered, it will be used as the default.
                Can also be provided as `device_id` query parameter.
            schema:
                type: string
                example: "my-device-id"
    securitySchemes:
        basicAuth:
            type: http
            scheme: basic
    schemas:
...
        SendResponse:
            type: object
            properties:
                code:
                    type: string
                    example: SUCCESS
                message:
                    type: string
                    example: Success
                results:
                    type: object
                    properties:
                        message_id:
                            type: string
                            example: "3EB0B430B6F8F1D0E053AC120E0A9E5C"
                        status:
                            type: string
                            example: "<feature> success ...."
...
        ErrorInternalServer:
            type: object
            properties:
                code:
                    type: string
                    example: INTERNAL_SERVER_ERROR
                    description: "SYSTEM_CODE_ERROR"
                message:
                    type: string
                    example: you are not loggin
                    description: "Detail error message"
                results:
                    type: object
                    example: null
                    description: "additional data"
        ErrorBadRequest:
            type: object
            properties:
                code:
                    type: string
                    example: 400
                    description: "HTTP Status Code"
                message:
                    type: string
                    example: field cannot be blank
                    description: "Detail error message"
                results:
                    type: object
                    example: null
                    description: "additional data"
...